requirements.txt
python-telegram-bot==20.7
python-decouple==3.8
Pillow==10.2.0
aiofiles==23.2.1
asyncio-mqtt==0.16.1
psutil==5.9.8
python-sane==2.9.1 